- Moved the text color definitions out of the executable and into an external   data file. - Removed the setcolor CCMD. It's been obsolete for years, ever since color-   aware cvars were added. - Changed V_GetColorStringByName() to return an FString, because it did a   copystring() call. - Extended V_GetColorFromString() so that it accepts HTML-style #RRGGBB and   #RGB color strings.
SVN r313 (trunk)

+++ src/textures/tgatexture.cpp
@@ -42,10 +42,9 @@
 
 bool FTGATexture::Check(FileReader & data)
 {
-	
 	TGAHeader hdr;
 	
-	if (data.GetLength() < sizeof(hdr)) return false;
+	if (data.GetLength() < (long)sizeof(hdr)) return false;
 	
 	data.Seek(0, SEEK_SET);
 	data.Read(&hdr, sizeof(hdr));
